+++ site/src/components/Workspace/Workspace.tsx
@@ -1,14 +1,10 @@
-import Box from "@material-ui/core/Box"
-import Paper from "@material-ui/core/Paper"
 import { makeStyles } from "@material-ui/core/styles"
 import Typography from "@material-ui/core/Typography"
-import CloudCircleIcon from "@material-ui/icons/CloudCircle"
 import React from "react"
-import { Link } from "react-router-dom"
 import * as Types from "../../api/types"
 import { WorkspaceSchedule } from "../WorkspaceSchedule/WorkspaceSchedule"
 import { WorkspaceSection } from "../WorkspaceSection/WorkspaceSection"
-import * as Constants from "./constants"
+import { WorkspaceStatusBar } from "../WorkspaceStatusBar/WorkspaceStatusBar"
 
 export interface WorkspaceProps {
   organization: Types.Organization
@@ -25,7 +21,7 @@ export const Workspace: React.FC<WorkspaceProps> = ({ organization, template, wo
   return (
     <div className={styles.root}>
       <div className={styles.vertical}>
-        <WorkspaceHeader organization={organization} template={template} workspace={workspace} />
+        <WorkspaceStatusBar organization={organization} template={template} workspace={workspace} />
         <div className={styles.horizontal}>
           <div className={styles.sidebarContainer}>
             <WorkspaceSection title="Applications">
@@ -56,40 +52,6 @@ export const Workspace: React.FC<WorkspaceProps> = ({ organization, template, wo
 }
 
 /**
- * Component for the header at the top of the workspace page
- */
-export const WorkspaceHeader: React.FC<WorkspaceProps> = ({ organization, template, workspace }) => {
-  const styles = useStyles()
-
-  const templateLink = `/templates/${organization.name}/${template.name}`
-
-  return (
-    <Paper elevation={0} className={styles.section}>
-      <div className={styles.horizontal}>
-        <WorkspaceHeroIcon />
-        <div className={styles.vertical}>
-          <Typography variant="h4">{workspace.name}</Typography>
-          <Typography variant="body2" color="textSecondary">
-            <Link to={templateLink}>{template.name}</Link>
-          </Typography>
-        </div>
-      </div>
-    </Paper>
-  )
-}
-
-/**
- * Component to render the 'Hero Icon' in the header of a workspace
- */
-export const WorkspaceHeroIcon: React.FC = () => {
-  return (
-    <Box mr="1em">
-      <CloudCircleIcon width={Constants.TitleIconSize} height={Constants.TitleIconSize} />
-    </Box>
-  )
-}
-
-/**
  * Temporary placeholder component until we have the sections implemented
  * Can be removed once the Workspace page has all the necessary sections
  */
@@ -115,12 +77,6 @@ export const useStyles = makeStyles((theme) => {
       display: "flex",
       flexDirection: "column",
     },
-    section: {
-      border: `1px solid ${theme.palette.divider}`,
-      borderRadius: Constants.CardRadius,
-      padding: Constants.CardPadding,
-      margin: theme.spacing(1),
-    },
     sidebarContainer: {
       display: "flex",
       flexDirection: "column",
@@ -128,10 +84,6 @@ export const useStyles = makeStyles((theme) => {
     },
     timelineContainer: {
       flex: 1,
-    },
-    icon: {
-      width: Constants.TitleIconSize,
-      height: Constants.TitleIconSize,
     },
   }
 })
