--- site/src/components/Workspace/constants.ts
@@ -1,3 +1,0 @@
-export const TitleIconSize = 48
-export const CardRadius = 8
-export const CardPadding = 20
