+++ site/src/theme/constants.ts
@@ -9,3 +9,6 @@ export const emptyBoxShadow = "none"
 export const navHeight = 56
 export const maxWidth = 1380
 export const sidePadding = "50px"
+export const TitleIconSize = 48
+export const CardRadius = 8
+export const CardPadding = 20
